test(musefs-fuse): extract shared FLAC/config fixtures into tests/common (closes #450)
The eight musefs-fuse integration tests each carried a byte-identical copy
of the FLAC fixture builders (flac_block/streaminfo_body/vorbis_comment_body/
make_flac) and the config()/MountConfig builder, with COVER_PNG +
flac_picture_block and walk_tree further duplicated across the ogg/playback
tests. Mirroring musefs-core/tests/common, add a tests/common/ module and
have every test use it.

The FLAC byte-layout primitives are reused from
musefs_format::fuzz_check::fixtures (the same source of truth musefs-core
uses), so the dev-dependency now enables the `fuzzing` feature. The two
stylistic variants (.unwrap() vs .expect(...)) produced identical bytes;
the centralized fixtures are byte-for-byte equivalent to the old copies, so
the audio invariant is unaffected. config()/config_with_mode(mode) replace
the per-file config()/mount_config()/core_config()/config(mode) spellings.

Net: 627 deletions for 37 insertions plus the shared module.
