+++ musefs-fuse/tests/concurrent_reads.rs
@@ -2,68 +2,13 @@
 //! and many files in parallel, driving the DbPool::PerThread worker pool.
 //! `--ignored` (needs /dev/fuse); runs in the e2e job and the TSan job.
 
-use std::collections::BTreeMap;
 use std::sync::{Arc, Barrier};
 
 use fuser::BackgroundSession;
-use musefs_core::{Mode, MountConfig, Musefs, scan_directory};
+use musefs_core::{Musefs, scan_directory};
 
-// --- Copy these helpers verbatim from mount.rs / concurrency.rs ---
-
-fn flac_block(block_type: u8, body: &[u8], is_last: bool) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.push((if is_last { 0x80 } else { 0 }) | (block_type & 0x7F));
-    let len = body.len();
-    out.push(u8::try_from((len >> 16) & 0xFF).unwrap());
-    out.push(u8::try_from((len >> 8) & 0xFF).unwrap());
-    out.push(u8::try_from(len & 0xFF).unwrap());
-    out.extend_from_slice(body);
-    out
-}
-
-fn streaminfo_body() -> Vec<u8> {
-    let mut b = vec![
-        0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0A, 0xC4, 0x42, 0xF0, 0x00,
-        0x00, 0x00, 0x00,
-    ];
-    b.extend_from_slice(&[0u8; 16]);
-    b
-}
-
-fn vorbis_comment_body(vendor: &str, comments: &[&str]) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(&u32::try_from(vendor.len()).unwrap().to_le_bytes());
-    out.extend_from_slice(vendor.as_bytes());
-    out.extend_from_slice(&u32::try_from(comments.len()).unwrap().to_le_bytes());
-    for c in comments {
-        out.extend_from_slice(&u32::try_from(c.len()).unwrap().to_le_bytes());
-        out.extend_from_slice(c.as_bytes());
-    }
-    out
-}
-
-fn make_flac(comments: &[&str], audio: &[u8]) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(b"fLaC");
-    out.extend_from_slice(&flac_block(0, &streaminfo_body(), false));
-    out.extend_from_slice(&flac_block(4, &vorbis_comment_body("orig", comments), true));
-    out.extend_from_slice(audio);
-    out
-}
-
-fn config() -> MountConfig {
-    MountConfig {
-        template: "$artist/$title".to_string(),
-        fallbacks: BTreeMap::new(),
-        default_fallback: "Unknown".to_string(),
-        mode: Mode::Synthesis,
-        poll_interval: std::time::Duration::ZERO,
-        case_insensitive: false,
-        read_ahead_budget: 64 * 1024 * 1024,
-        read_ahead_prefetch: false,
-        skip_on_missing: false,
-    }
-}
+mod common;
+use common::{config, make_flac};
 
 // ---------------------------------------------------------------------------
