+++ musefs-fuse/tests/keep_cache.rs
@@ -1,66 +1,9 @@
-use std::collections::BTreeMap;
 use std::time::Duration;
 
-use musefs_core::{MountConfig, Musefs, scan_directory};
+use musefs_core::{Musefs, scan_directory};
 
-// --- minimal proven FLAC fixture (mirrors musefs-fuse/tests/mount.rs) ---
-
-fn flac_block(block_type: u8, body: &[u8], is_last: bool) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.push((if is_last { 0x80 } else { 0 }) | (block_type & 0x7F));
-    let len = body.len();
-    out.push(u8::try_from((len >> 16) & 0xFF).unwrap());
-    out.push(u8::try_from((len >> 8) & 0xFF).unwrap());
-    out.push(u8::try_from(len & 0xFF).unwrap());
-    out.extend_from_slice(body);
-    out
-}
-
-fn streaminfo_body() -> Vec<u8> {
-    let mut b = vec![
-        0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0A, 0xC4, 0x42, 0xF0, 0x00,
-        0x00, 0x00, 0x00,
-    ];
-    b.extend_from_slice(&[0u8; 16]);
-    b
-}
-
-fn vorbis_comment_body(vendor: &str, comments: &[&str]) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(&u32::try_from(vendor.len()).unwrap().to_le_bytes());
-    out.extend_from_slice(vendor.as_bytes());
-    out.extend_from_slice(&u32::try_from(comments.len()).unwrap().to_le_bytes());
-    for c in comments {
-        out.extend_from_slice(&u32::try_from(c.len()).unwrap().to_le_bytes());
-        out.extend_from_slice(c.as_bytes());
-    }
-    out
-}
-
-fn make_flac(comments: &[&str], audio: &[u8]) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(b"fLaC");
-    out.extend_from_slice(&flac_block(0, &streaminfo_body(), false));
-    out.extend_from_slice(&flac_block(4, &vorbis_comment_body("orig", comments), true));
-    out.extend_from_slice(audio);
-    out
-}
-
-/// Scan a backing dir into a Db with poll_interval=ZERO so metadata ops
-/// trigger refresh immediately.
-fn mount_config() -> MountConfig {
-    MountConfig {
-        template: "$artist/$title".to_string(),
-        fallbacks: BTreeMap::new(),
-        default_fallback: "Unknown".to_string(),
-        mode: musefs_core::Mode::Synthesis,
-        poll_interval: Duration::ZERO,
-        case_insensitive: false,
-        read_ahead_budget: 64 * 1024 * 1024,
-        read_ahead_prefetch: false,
-        skip_on_missing: false,
-    }
-}
+mod common;
+use common::{config, make_flac};
 
 #[test]
 #[ignore = "requires /dev/fuse; run with: cargo test -p musefs-fuse -- --ignored"]
@@ -76,7 +19,7 @@ fn keep_cache_mount_reflects_retag_after_refresh() {
     let db = musefs_db::Db::open(&db_path).unwrap();
     scan_directory(&db, backing.path()).unwrap();
 
-    let fs = Musefs::open(db, mount_config()).unwrap();
+    let fs = Musefs::open(db, config()).unwrap();
 
     // 2. Mount with keep-cache=true and debounce=ZERO (already set via MountConfig).
     let cfg = musefs_fuse::FuseConfig {
