+++ musefs-fuse/tests/ogg_read_through.rs
@@ -1,7 +1,8 @@
-use std::collections::BTreeMap;
-
 use base64::Engine as _;
-use musefs_core::{MountConfig, Musefs, scan_directory};
+use musefs_core::{Musefs, scan_directory};
+
+mod common;
+use common::{COVER_PNG, config, flac_picture_block};
 
 /// Encode a tiny tagged fixture with ffmpeg. `args` are the codec-specific ffmpeg
 /// args (codec + container). Returns false (skip) if ffmpeg/codec is unavailable.
@@ -68,20 +69,6 @@ fn find_one_file(root: &std::path::Path) -> std::path::PathBuf {
     }
 }
 
-fn config() -> MountConfig {
-    MountConfig {
-        template: "$artist/$title".to_string(),
-        fallbacks: BTreeMap::new(),
-        default_fallback: "Unknown".to_string(),
-        mode: musefs_core::Mode::Synthesis,
-        poll_interval: std::time::Duration::ZERO,
-        case_insensitive: false,
-        read_ahead_budget: 64 * 1024 * 1024,
-        read_ahead_prefetch: false,
-        skip_on_missing: false,
-    }
-}
-
 /// Mount a single-track backing dir containing `src`, read the one synthesized
 /// file back, and validate: all page CRCs valid (read_packets), the comment packet
 /// carries the rewritten title, and the AUDIO packets are byte-identical to the
@@ -130,37 +117,6 @@ fn mount_and_validate(src: &std::path::Path) {
     );
 
     drop(session);
-}
-
-/// A valid 4x4 PNG cover image. ffmpeg 8's PNG decoder rejects malformed chunks,
-/// so this must be a real, decodable image.
-const COVER_PNG: &[u8] = &[
-    0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A, 0x00, 0x00, 0x00, 0x0D, 0x49, 0x48, 0x44, 0x52,
-    0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x08, 0x02, 0x00, 0x00, 0x00, 0x26, 0x93, 0x09,
-    0x29, 0x00, 0x00, 0x00, 0x09, 0x70, 0x48, 0x59, 0x73, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
-    0x01, 0x00, 0x4F, 0x25, 0xC4, 0xD6, 0x00, 0x00, 0x00, 0x14, 0x49, 0x44, 0x41, 0x54, 0x78, 0x9C,
-    0x63, 0x64, 0x60, 0xF8, 0xC7, 0x00, 0x03, 0x2C, 0x0C, 0x48, 0x00, 0x37, 0x07, 0x00, 0x32, 0x3E,
-    0x01, 0x0C, 0x1C, 0xDB, 0xAF, 0x41, 0x00, 0x00, 0x00, 0x00, 0x49, 0x45, 0x4E, 0x44, 0xAE, 0x42,
-    0x60, 0x82,
-];
-
-/// Build a FLAC METADATA PICTURE block body: picture type, MIME, description,
-/// dimensions, then the image. Base64-encoded, this is a Vorbis
-/// `METADATA_BLOCK_PICTURE` comment value. Big-endian.
-fn flac_picture_block(png: &[u8]) -> Vec<u8> {
-    let mime: &[u8] = b"image/png";
-    let mut out = Vec::new();
-    out.extend_from_slice(&3u32.to_be_bytes()); // type: front cover
-    out.extend_from_slice(&u32::try_from(mime.len()).unwrap().to_be_bytes());
-    out.extend_from_slice(mime);
-    out.extend_from_slice(&0u32.to_be_bytes()); // description length (empty)
-    out.extend_from_slice(&4u32.to_be_bytes()); // width
-    out.extend_from_slice(&4u32.to_be_bytes()); // height
-    out.extend_from_slice(&24u32.to_be_bytes()); // color depth
-    out.extend_from_slice(&0u32.to_be_bytes()); // colors used (0 = non-indexed)
-    out.extend_from_slice(&u32::try_from(png.len()).unwrap().to_be_bytes());
-    out.extend_from_slice(png);
-    out
 }
 
 /// Generate an Ogg fixture carrying a cover image via a base64
