+++ musefs-fuse/tests/passthrough.rs
@@ -7,81 +7,12 @@
 //! Run with:
 //!   cargo test -p musefs-fuse --features metrics --test passthrough -- --ignored --nocapture --test-threads=1
 
-use std::collections::BTreeMap;
 use std::io::Read;
 
-use musefs_core::{Mode, MountConfig, Musefs, metrics, scan_directory};
+use musefs_core::{Mode, Musefs, metrics, scan_directory};
 
-// ---------------------------------------------------------------------------
-// Minimal proven FLAC fixture (mirrors tests/mount.rs exactly)
-// ---------------------------------------------------------------------------
-
-fn flac_block(block_type: u8, body: &[u8], is_last: bool) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.push((if is_last { 0x80 } else { 0 }) | (block_type & 0x7F));
-    let len = body.len();
-    out.push(u8::try_from((len >> 16) & 0xFF).expect("FLAC block length high byte fits in u8"));
-    out.push(u8::try_from((len >> 8) & 0xFF).expect("FLAC block length middle byte fits in u8"));
-    out.push(u8::try_from(len & 0xFF).expect("FLAC block length low byte fits in u8"));
-    out.extend_from_slice(body);
-    out
-}
-
-fn streaminfo_body() -> Vec<u8> {
-    let mut b = vec![
-        0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0A, 0xC4, 0x42, 0xF0, 0x00,
-        0x00, 0x00, 0x00,
-    ];
-    b.extend_from_slice(&[0u8; 16]);
-    b
-}
-
-fn vorbis_comment_body(vendor: &str, comments: &[&str]) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(
-        &u32::try_from(vendor.len())
-            .expect("vendor length fits in u32")
-            .to_le_bytes(),
-    );
-    out.extend_from_slice(vendor.as_bytes());
-    out.extend_from_slice(
-        &u32::try_from(comments.len())
-            .expect("comment count fits in u32")
-            .to_le_bytes(),
-    );
-    for c in comments {
-        out.extend_from_slice(
-            &u32::try_from(c.len())
-                .expect("comment length fits in u32")
-                .to_le_bytes(),
-        );
-        out.extend_from_slice(c.as_bytes());
-    }
-    out
-}
-
-fn make_flac(comments: &[&str], audio: &[u8]) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(b"fLaC");
-    out.extend_from_slice(&flac_block(0, &streaminfo_body(), false));
-    out.extend_from_slice(&flac_block(4, &vorbis_comment_body("orig", comments), true));
-    out.extend_from_slice(audio);
-    out
-}
-
-fn config(mode: Mode) -> MountConfig {
-    MountConfig {
-        template: "$artist/$title".to_string(),
-        fallbacks: BTreeMap::new(),
-        default_fallback: "Unknown".to_string(),
-        mode,
-        poll_interval: std::time::Duration::ZERO,
-        case_insensitive: false,
-        read_ahead_budget: 64 * 1024 * 1024,
-        read_ahead_prefetch: false,
-        skip_on_missing: false,
-    }
-}
+mod common;
+use common::{config_with_mode, make_flac};
 
 /// FUSE passthrough landed in mainline 6.9.
 fn kernel_supports_passthrough() -> bool {
@@ -114,7 +45,7 @@ fn mount_one_track(
     let db = musefs_db::Db::open(&db_path).unwrap();
     scan_directory(&db, backing.path()).unwrap();
 
-    let fs = Musefs::open(db, config(mode)).unwrap();
+    let fs = Musefs::open(db, config_with_mode(mode)).unwrap();
     let mountpoint = tempfile::tempdir().unwrap();
     let session = musefs_fuse::spawn(fs, mountpoint.path(), "musefs-passthrough-test").unwrap();
     let mnt = mountpoint.keep(); // keep mount alive for the test's duration
